A raw version of the microservice Events.API completed

+++ Events.API/Events.Infrastructure/Services/EventService.cs
@@ -14,26 +14,42 @@ public class EventService : IEventService
 
     public async Task<int> AddEvent(Event @event)
     {
+        if (@event == null) throw new ArgumentNullException(nameof(@event));
+        if (!IsValidEvent(@event)) throw new ArgumentException("Invalid event data.");
+
         return await _eventRepository.AddAsync(@event) ?? 0;
     }
 
     public async Task DeleteEvent(int id)
     {
+        if (id <= 0) throw new ArgumentException("Invalid event ID.");
         await _eventRepository.DeleteByIdAsync(id);
     }
 
     public async Task<Event> GetEvent(int id)
     {
-        return await _eventRepository.GetByIdAsync(id) ?? new Event();
+        if (id <= 0) throw new ArgumentException("Invalid event ID.");
+        var @event = await _eventRepository.GetByIdAsync(includeProperties: "Polls,EventParticipants", ids: id);
+        if (@event == null) throw new KeyNotFoundException("Event not found.");
+        return @event;
     }
 
     public async Task<IEnumerable<Event>> GetEvents()
     {
-        return await _eventRepository.Get();
+        return await _eventRepository.Get(includeProperties: "Polls,EventParticipants");
     }
 
-    public Task<bool> UpdateEvent(Event @event)
+    public async Task<bool> UpdateEvent(Event @event)
     {
-        return _eventRepository.UpdateAsync(@event);
+        if (@event == null) throw new ArgumentNullException(nameof(@event));
+        if (!IsValidEvent(@event)) throw new ArgumentException("Invalid event data.");
+
+        return await _eventRepository.UpdateAsync(@event);
+    }
+
+    private bool IsValidEvent(Event @event)
+    {
+        // Здесь должна быть логика проверки полей (нужно будет создать класс валидатор через fluent validation)
+        return @event.Name != null && @event.Time != null;
     }
 }
